+++ ft_memcpy.c
@@ -9,10 +9,11 @@
 /*   Updated: 2025/04/09 13:00:09 by ejavier-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+#include "libft.h"
 
-void	*ft_memcpy(void *dest, const void *src, unsigned int n)
+void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 	char	*d;
 	char	*s;
